refactor: modularized side panel
